+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed overwriting STARTING job status in a sync loop.
+  ([#457](https://github.com/TheRacetrack/racetrack/issues/457))
 
 ## [2.29.1] - 2024-04-26
 ### Fixed